chore: fix import paths after refactor
