+++ bciers/apps/reporting/src/app/bceidbusiness/industry_user_admin/reports/[version_id]/facilities/[facility_id]/activities/gsc_solely_for_line_tracing/page.tsx
@@ -1,4 +1,4 @@
-import GeneralStationaryCombustion from "../../../../components/activities/generalStationaryCombustion";
+import GeneralStationaryCombustion from "../../../../../../../../components/activities/generalStationaryCombustion";
 import { Suspense } from "react";
 import { actionHandler } from "@bciers/actions";
 import safeJsonParse from "libs/utils/safeJsonParse";
